dj_model/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
celery-with-redis==3.0
chardet==3.0.4
click==7.1.1
cycler==0.10.0
diff-match-patch==20181111
Django==3.0.5
django-celery==3.3.1
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==2.0.2
django-jsonfield==1.4.0
django-mdeditor==0.1.18
django-redis==4.10.0
django-reversion==3.0.4
django-uuslug==1.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
importlib-metadata==1.6.0
kombu==4.6.8
mysqlclient==1.4.6
py3Fdfs==2.2.0
PyJWT==1.7.1
PyMySQL==0.9.3
python-slugify==4.0.0
pytz==2020.1
redis==3.5.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
vine==1.3.0
zipp==3.1.0

APScheduler~=3.6.3
django-debug-toolbar~=2.2